+++ packages/server/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.2-alpha.3](https://github.com/niledatabase/nile-js/compare/v4.0.2-alpha.2...v4.0.2-alpha.3) (2025-04-11)
+
+**Note:** Version bump only for package @niledatabase/server
+
+## [4.0.2-alpha.2](https://github.com/niledatabase/nile-js/compare/v4.0.1...v4.0.2-alpha.2) (2025-04-11)
+
+### Bug Fixes
+
+- **server:** add util function for returning the auth cookie ([d62d4f0](https://github.com/niledatabase/nile-js/commit/d62d4f040789f5588a153738011552ec3e398738))
+- **server:** update server side create user with correct params ([2dba5dd](https://github.com/niledatabase/nile-js/commit/2dba5ddf9d55989f48455d4de67d0c657da8de37))
+
+## [4.0.2-alpha.1](https://github.com/niledatabase/nile-js/compare/v4.0.2-alpha.0...v4.0.2-alpha.1) (2025-04-11)
+
+**Note:** Version bump only for package @niledatabase/server
+
+## [4.0.2-alpha.0](https://github.com/niledatabase/nile-js/compare/v4.0.1...v4.0.2-alpha.0) (2025-04-11)
+
+### Bug Fixes
+
+- **server:** add util function for returning the auth cookie ([d62d4f0](https://github.com/niledatabase/nile-js/commit/d62d4f040789f5588a153738011552ec3e398738))
+- **server:** update server side create user with correct params ([2dba5dd](https://github.com/niledatabase/nile-js/commit/2dba5ddf9d55989f48455d4de67d0c657da8de37))
+
 ## [4.0.1](https://github.com/niledatabase/nile-js/compare/v4.0.1-alpha.1...v4.0.1) (2025-04-09)
 
 **Note:** Version bump only for package @niledatabase/server